LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-sizing-expected.txt


FAIL flip-block does not affect sizing assert_equals: expected "1px" but got "0px"
FAIL flip-inline does not affect sizing assert_equals: expected "1px" but got "0px"
FAIL flip-start affects sizing assert_equals: expected "40px" but got "30px"
